refactor *_config into config
