+++ config/metaclip_data.py
@@ -22,3 +22,4 @@ class metaclip_2_5b:
     start_shard = 0
     end_shard = 200000
     max_match = 170000
+
